attendance_widget_release/ubuntu/app/pyproject.toml
[project]
name = "attendance-widget-app"
version = "0.1.0"
description = "Desktop widget for attendance automation and weekly work balance."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "PySide6>=6.9.0",
    "python-dotenv>=1.1.0",
    "selenium>=4.31.0",
]

[project.scripts]
attendance-widget = "attendance_widget.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/attendance_widget"]

[tool.uv]
package = true
